Cleanup folder structure in rosbag2_storage and rosbag2_tests
- use src/<package_name>/ and test/<package_name>/ folders everywhere
- harmonises with all other packages
- results in better header guards
