+++ rosbag2_storage/test/rosbag2_storage/test_read_only_plugin.cpp
