+++ rosbag2_tests/test/rosbag2_tests/subscription_manager.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SUBSCRIPTION_MANAGER_HPP_
-#define SUBSCRIPTION_MANAGER_HPP_
+#ifndef ROSBAG2_TESTS__SUBSCRIPTION_MANAGER_HPP_
+#define ROSBAG2_TESTS__SUBSCRIPTION_MANAGER_HPP_
 
 #include <future>
 #include <map>
@@ -96,4 +96,4 @@ private:
 
 }  // namespace test_helpers
 
-#endif  // SUBSCRIPTION_MANAGER_HPP_
+#endif  // ROSBAG2_TESTS__SUBSCRIPTION_MANAGER_HPP_
