+++ rosbag2_tests/test/rosbag2_tests/temporary_directory_fixture.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TEMPORARY_DIRECTORY_FIXTURE_HPP_
-#define TEMPORARY_DIRECTORY_FIXTURE_HPP_
+#ifndef ROSBAG2_TESTS__TEMPORARY_DIRECTORY_FIXTURE_HPP_
+#define ROSBAG2_TESTS__TEMPORARY_DIRECTORY_FIXTURE_HPP_
 
 #include <gmock/gmock.h>
 
@@ -100,4 +100,4 @@ public:
 
 std::string TemporaryDirectoryFixture::temporary_dir_path_ = "";  // NOLINT
 
-#endif  // TEMPORARY_DIRECTORY_FIXTURE_HPP_
+#endif  // ROSBAG2_TESTS__TEMPORARY_DIRECTORY_FIXTURE_HPP_
